// src/main/models/entity/master/MasterMenuItems.ts
import {Sequelize} from 'sequelize';
import {Column, DataType, Model, Table} from 'sequelize-typescript';

export interface MasterMenuItemsAttributes {
  id: number;
  icon?: string;
  labelKey: string;
  label: string;
  urlPath?: string;
  parentId?: number;
  indexed?:number;
  flag?: number;
  createdBy: string;
  createdAt: Date;
  updatedBy?: string;
  updatedAt?: Date;
  deletedAt?: Date;
}

type MasterMenuItemsAutoGeneratedAttributes =
  | 'id'
  | 'createdAt'
  | 'updatedAt'
  | 'deletedAt';
export type MasterMenuItemsCreationAttributes = Omit<
  MasterMenuItemsAttributes,
  MasterMenuItemsAutoGeneratedAttributes
>;

@Table({
  freezeTableName: true,
  tableName: 'menu_items',
  underscored: true,
  paranoid: true,
  schema: 'master',
})
class MasterMenuItems
  extends Model<MasterMenuItemsAttributes | MasterMenuItemsCreationAttributes>
  implements MasterMenuItemsAttributes
{
  @Column({
    allowNull: false,
    type: DataType.BIGINT,
    primaryKey: true,
    autoIncrement: true,
  })
  id!: number;

  @Column({
    allowNull: true,
    type: DataType.STRING,
  })
  icon!: string;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  labelKey!: string;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  label!: string;

  @Column({
    allowNull: true,
    type: DataType.TEXT,
  })
  urlPath!: string;

  @Column({
    allowNull: true,
    type: DataType.BIGINT,
    defaultValue:0
  })
  parentId!: number;

  @Column({
    allowNull: true,
    type: DataType.BIGINT,
    defaultValue:0
  })
  indexed!: number;

  @Column({
    allowNull: true,
    type: DataType.INTEGER,
    defaultValue: 1,
  })
  flag!: number;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  createdBy!: string;

  @Column({
    allowNull: false,
    type: DataType.DATE,
    defaultValue: Sequelize.literal('now()'),
  })
  createdAt!: Date;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  updatedBy!: string;

  @Column({
    type: DataType.DATE,
  })
  updatedAt!: Date;

  @Column({
    type: DataType.DATE,
  })
  deletedAt!: Date;
}

export default MasterMenuItems;